# src/api/swagger.yaml
basePath: /api/v1
definitions:
...
  models.AuthProviderType:
    enum:
    - basic
    - ldap
    - oidc
    type: string
    x-enum-varnames:
    - AuthProviderBasic
    - AuthProviderLDAP
    - AuthProviderOIDC
...
  types.AuthProviderAddDto:
    properties:
      display_name:
        type: string
      enabled:
        type: boolean
      ldap_base_dn:
        type: string
      ldap_bind_dn:
        type: string
      ldap_bind_pass:
        type: string
      ldap_filter:
        type: string
      ldap_tls_cert:
        type: string
      ldap_url:
        description: LDAP fields
        type: string
      name:
        type: string
      oidc_client_id:
        type: string
      oidc_client_secret:
        type: string
      oidc_issuer_url:
        description: OIDC fields
        type: string
      oidc_redirect_url:
        type: string
      priority:
        type: integer
      type:
        $ref: '#/definitions/models.AuthProviderType'
    required:
    - name
    - type
    type: object
  types.AuthProviderUpdateDto:
    properties:
      display_name:
        type: string
      enabled:
        type: boolean
      id:
        type: integer
      ldap_base_dn:
        type: string
      ldap_bind_dn:
        type: string
      ldap_bind_pass:
        description: Only update if not empty
        type: string
      ldap_filter:
        type: string
      ldap_tls_cert:
        type: string
      ldap_url:
        description: LDAP fields
        type: string
      name:
        type: string
      oidc_client_id:
        type: string
      oidc_client_secret:
        description: Only update if not empty
        type: string
      oidc_issuer_url:
        description: OIDC fields
        type: string
      oidc_redirect_url:
        type: string
      priority:
        type: integer
      type:
        $ref: '#/definitions/models.AuthProviderType'
    required:
    - id
    - name
    - type
    type: object
...
  types.LdapLoginDto:
    properties:
      password:
        type: string
      provider:
        description: Provider name
        type: string
      username:
        type: string
    required:
    - password
    - provider
    - username
    type: object
  types.LoginDto:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  types.OidcCallbackDto:
    properties:
      code:
        type: string
      state:
        type: string
    required:
    - code
    - state
    type: object
...
  types.UpdateTokenDto:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
...
host: localhost:8086
info:
  contact:
    name: Antelope Team
  description: Antelope bioinformatics pipeline management platform API.
  license:
    name: MIT
  title: Antelope API
  version: "1.0"
paths:
...
  /auth/ldap/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user via LDAP
      parameters:
      - description: LDAP credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.LdapLoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: LDAP login
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Local login
      tags:
      - auth
  /auth/logout:
    post:
      description: Log out the current user and invalidate the session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /auth/oidc/{provider}/start:
    get:
      description: Mint a one-time state nonce and return the provider authorization
        URL to redirect to
      parameters:
      - description: Provider name
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Start OIDC login
      tags:
      - auth
  /auth/oidc/callback:
    post:
      consumes:
      - application/json
      description: Handle the OIDC provider callback and complete authentication
      parameters:
      - description: OIDC callback payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.OidcCallbackDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: OIDC callback
      tags:
      - auth
  /auth/providers:
    get:
      description: Get the list of currently enabled authentication providers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: List enabled auth providers
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Obtain a new access token using a refresh token
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.UpdateTokenDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Refresh access token
      tags:
      - auth
...
  /system/auth/provider/add:
    post:
      consumes:
      - application/json
      description: Add a new authentication provider
      parameters:
      - description: Auth provider config
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.AuthProviderAddDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add auth provider
      tags:
      - auth-providers
  /system/auth/provider/delete/{id}:
    delete:
      description: Delete an authentication provider by ID
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete auth provider
      tags:
      - auth-providers
  /system/auth/provider/test/{id}:
    post:
      description: Test the connectivity/configuration of an authentication provider
        by ID
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Test auth provider
      tags:
      - auth-providers
  /system/auth/provider/update:
    post:
      consumes:
      - application/json
      description: Update an existing authentication provider
      parameters:
      - description: Auth provider config
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.AuthProviderUpdateDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update auth provider
      tags:
      - auth-providers
  /system/auth/providers:
    get:
      description: Get the list of all configured authentication providers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all auth providers
      tags:
      - auth-providers
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {accessToken}"
    in: header
    name: Authorization
    type: apiKey
  RefreshAuth:
    description: Enter "Bearer {refreshToken}" — used only for /auth/refresh
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"